requirements.txt
librosa==0.9.2
numpy==1.23.5
PyYAML==5.4.1
scikit_learn==1.0.2
scipy==1.12.0
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.66.1
